+++ lib/dart_quill_table_better.dart
@@ -5,3 +5,4 @@ export 'src/table_better/formats/table.dart';
 export 'src/table_better/language/language.dart';
 export 'src/table_better/modules/clipboard.dart';
 export 'src/table_better/register.dart';
+export 'src/table_better/ui/cell_selection.dart';